+++ packages/toolbar/src/core/context/ReactMountContext.tsx
@@ -1,5 +1,9 @@
-import { createContext } from 'react';
+import { createContext, useContext } from 'react';
 
 const ReactMountContext = createContext<HTMLElement>(document.createElement('div'));
+
+export const useReactMount = () => {
+  return useContext(ReactMountContext);
+};
 
 export default ReactMountContext;